+++ src/OpenCV/FaceRecognizer.cpp
@@ -44,8 +44,8 @@ cv::CascadeClassifier OpenCV::FaceRecognizer::getCascadeClassifier()
 
 void OpenCV::FaceRecognizer::computeEyesCoordinations(Rect face, Size size)
 {
-	this->eyesCoord.x = (((float)(face.x+face.width/2) / (float)size.width-0.5)/0.5)*100;
-	this->eyesCoord.y = (((float)(face.y+face.height/3) / (float)size.height-0.5)/0.5)*100;
+	this->eyesCoord.x = (((float)(face.x+face.width/2) / (float)size.width-0.5f)/0.5f)*100;
+	this->eyesCoord.y = (((float)(face.y+face.height/3) / (float)size.height-0.5f)/0.5f)*100;
 }
 
 cv::Point2i OpenCV::FaceRecognizer::getEyesCoords()
@@ -53,7 +53,7 @@ cv::Point2i OpenCV::FaceRecognizer::getEyesCoords()
 	return this->eyesCoord;
 }
 
-float OpenCV::FaceRecognizer::getHeadDistance(int screenWidth)
+float OpenCV::FaceRecognizer::getHeadDistance(double screenWidth)
 {
 	if (this->faces.size()>0)
 	{
